[UI테스트][4단계] UI 계산기

+++ domain/src/test/java/edu/nextstep/camp/calculator/domain/OperationTest.kt
@@ -5,30 +5,32 @@ import org.junit.Test
 import org.junit.runner.RunWith
 import org.junit.runners.Parameterized
 import org.junit.runners.Parameterized.Parameters
+import java.math.BigInteger
 
 @RunWith(Parameterized::class)
 class OperationTest(
     operator: Char,
     private val operation: Operation,
-    private val left: Int,
-    private val right: Int,
-    private val expected: Int
+    private val left: BigInteger,
+    private val right: BigInteger,
+    private val expected: BigInteger
 ) {
 
     companion object {
         @JvmStatic
         @Parameters(name = " {2} {0} {3} = {4} ")
         fun getTestParameters() = listOf(
-            arrayOf(Operation.Plus.operator, Operation.Plus, 2, 3, 5),
-            arrayOf(Operation.Minus.operator, Operation.Minus, 2, 3, -1),
-            arrayOf(Operation.Div.operator,  Operation.Div, 6, 3, 2),
-            arrayOf(Operation.Mult.operator, Operation.Mult, 2, 3, 6),
+            arrayOf(Operation.Plus.operator, Operation.Plus, 2.toBigInteger(), 3.toBigInteger(), 5.toBigInteger()),
+            arrayOf(Operation.Minus.operator, Operation.Minus, 2.toBigInteger(), 3.toBigInteger(), (-1).toBigInteger()),
+            arrayOf(Operation.Div.operator,  Operation.Div, 6.toBigInteger(), 3.toBigInteger(), 2.toBigInteger()),
+            arrayOf(Operation.Mult.operator, Operation.Mult, 2.toBigInteger(), 3.toBigInteger(), 6.toBigInteger()),
         )
     }
 
     @Test
     fun `연산`() {
         //when
+
         val actual = operation(left, right)
 
         //then
